third_party/blink/web_tests/external/wpt/css/css-transforms/parsing/transform-box-computed-expected.txt

This is a testharness.js-based test.
FAIL Property transform-box value 'content-box' computes to 'content-box' assert_equals: expected "content-box" but got "view-box"
FAIL Property transform-box value 'border-box' computes to 'border-box' assert_equals: expected "border-box" but got "view-box"
PASS Property transform-box value 'fill-box' computes to 'fill-box'
FAIL Property transform-box value 'stroke-box' computes to 'stroke-box' assert_equals: expected "stroke-box" but got "view-box"
PASS Property transform-box value 'view-box' computes to 'view-box'
Harness: the test ran to completion.
